refactored the UDP transport to push most of the code and logic back into the transport layer itself.
* Command how has transient from & to properties which can be used with transports like UDP/multicast to indicate which endpoint (typically broker) actually sent the commands
* used int for commandId 
* support for PartialCommand support; allowing large commands (such as big messages) to be split up into smaller chunks
* added a CommandJoiner for joining partial commands together into complete commands
* ReliableTransport which re-orders and can re-request missed commands


git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/trunk@384826 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-core/src/main/java/org/apache/activemq/broker/AbstractConnection.java
@@ -186,7 +186,7 @@ public abstract class AbstractConnection implements Service, Connection, Task, C
         
         Response response=null;
         boolean responseRequired = command.isResponseRequired();
-        short commandId = command.getCommandId();
+        int commandId = command.getCommandId();
         try {
             response = command.visit(this);
         } catch ( Throwable e ) {
